+++ src/main/java/com/cotato/kampus/domain/chat/application/PostChatService.java
@@ -18,6 +18,22 @@ import com.cotato.kampus.domain.chat.domain.ChatNotificationResult;
 import com.cotato.kampus.domain.chat.domain.ChatRoomDetailDto;
 import com.cotato.kampus.domain.chat.domain.ChatRoomPreview;
 import com.cotato.kampus.domain.chat.domain.ChatRoomPreviewList;
+import com.cotato.kampus.domain.chat.implement.ChatMemberFinder;
+import com.cotato.kampus.domain.chat.implement.message.ChatMessageAppender;
+import com.cotato.kampus.domain.chat.implement.message.ChatMessageDeleter;
+import com.cotato.kampus.domain.chat.implement.message.ChatMessageFinder;
+import com.cotato.kampus.domain.chat.implement.message.ChatMessageProcessor;
+import com.cotato.kampus.domain.chat.implement.chatroom.ChatRoomAppender;
+import com.cotato.kampus.domain.chat.implement.chatroom.ChatRoomDeleter;
+import com.cotato.kampus.domain.chat.implement.chatroom.ChatRoomFinder;
+import com.cotato.kampus.domain.chat.implement.chatroom.ChatRoomValidator;
+import com.cotato.kampus.domain.chat.implement.metadata.ChatroomMetadataAppender;
+import com.cotato.kampus.domain.chat.implement.metadata.ChatroomMetadataDeleter;
+import com.cotato.kampus.domain.chat.implement.metadata.ChatroomMetadataFinder;
+import com.cotato.kampus.domain.chat.implement.metadata.ChatroomMetadataMapper;
+import com.cotato.kampus.domain.chat.implement.metadata.ChatroomMetadataUpdater;
+import com.cotato.kampus.domain.chat.implement.read.MessageReadStatusDeleter;
+import com.cotato.kampus.domain.chat.implement.read.MessageReadStatusUpdater;
 import com.cotato.kampus.domain.common.application.ApiUserResolver;
 import com.cotato.kampus.domain.post.application.PostFinder;
 import com.cotato.kampus.domain.post.dto.PostDto;
