+++ src/main/java/com/cotato/kampus/domain/chat/implement/metadata/ChatroomMetadataAppender.java
@@ -1,4 +1,4 @@
-package com.cotato.kampus.domain.chat.application;
+package com.cotato.kampus.domain.chat.implement.metadata;
 
 import java.time.LocalDateTime;
 import java.util.List;
