+++ src/main/java/com/cotato/kampus/domain/chat/implement/metadata/ChatroomMetadataUpdater.java
@@ -1,4 +1,4 @@
-package com.cotato.kampus.domain.chat.application;
+package com.cotato.kampus.domain.chat.implement.metadata;
 
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
