+++ src/main/java/com/cotato/kampus/domain/chat/implement/read/MessageReadStatusDeleter.java
@@ -1,9 +1,9 @@
-package com.cotato.kampus.domain.chat.application;
+package com.cotato.kampus.domain.chat.implement.read;
 
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.cotato.kampus.domain.chat.dao.repository.ChatMessageRepository;
+import com.cotato.kampus.domain.chat.dao.repository.MessageReadStatusRepository;
 
 import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
@@ -11,11 +11,11 @@ import lombok.RequiredArgsConstructor;
 @Component
 @Transactional(readOnly = true)
 @RequiredArgsConstructor(access = AccessLevel.PROTECTED)
-public class ChatMessageDeleter {
-	private final ChatMessageRepository chatMessageRepository;
+public class MessageReadStatusDeleter {
+	private final MessageReadStatusRepository messageReadStatusRepository;
 
 	@Transactional
 	public void deleteByChatroomId(Long chatroomId) {
-		chatMessageRepository.deleteAllByChatroomId(chatroomId);
+		messageReadStatusRepository.deleteAllByChatroomId(chatroomId);
 	}
 }
